[DMG] Move DMG07 related stuff to its own CPP file
